Rework for new kernel patches
Update fsverity-utils for the new design where the Merkle tree is built
in the kernel by FS_IOC_ENABLE_VERITY.

'fsverity enable' command is updated to pass the new struct to the
ioctl, and to accept the Merkle tree parameters on the command line.

'fsverity sign' command is added.  It calculates a file's fs-verity
measurement and signs it.  Previously this functionality was part of
'fsverity setup'.

'fsverity setup' command is removed.

Signed-off-by: Eric Biggers <ebiggers@google.com>