cmd_sign: create detached signatures
There doesn't appear to be a good reason to include the file measurement
in the signatures themselves, and it makes the kernel-side verification
more complex since we have to manually compare the actual file
measurement to the actual signed data in the PKCS#7 message.  So let's
switch to using detached signatures instead.

Signed-off-by: Eric Biggers <ebiggers@google.com>