Move libfsverity.h to its own directory
libfsverity.h is the public API, but the other headers in common/ are
private headers for fsverity-utils.  Move libfsverity.h to its own
directory to make this clear.  This is also needed for Android's build
system in order to restrict the exported headers to libfsverity.h.

This doesn't affect users who are using 'make install', since
'make install' still installs libfsverity.h to the same place,
and it doesn't install any private headers.

Link: https://lore.kernel.org/r/20200923202328.16310-1-ebiggers@kernel.org
Signed-off-by: Eric Biggers <ebiggers@google.com>